Move effect handling from Skill to Effect
