+++ apps/arena/lib/arena/game/skill.ex
@@ -295,124 +295,6 @@ defmodule Arena.Game.Skill do
     end)
   end
 
-  ## TODO: refactor into Effect module, take a closer look at how the recurring effects are re-applied
-  ##    specifically the mess around apply_effect_mechanic/3 and do-effect_mechanics/4
-  def apply_effect_mechanic(%{players: players, pools: pools} = game_state) do
-    game_state =
-      Enum.reduce(players, game_state, fn {_player_id, player}, game_state ->
-        if Player.alive?(player) do
-          player =
-            Enum.reduce(player.aditional_info.effects, player, fn effect, player ->
-              apply_effect_mechanic(player, effect, game_state)
-            end)
-
-          put_in(game_state, [:players, player.id], player)
-        else
-          game_state
-        end
-      end)
-
-    Enum.reduce(pools, game_state, fn {_pool_id, pool}, game_state ->
-      pool =
-        Enum.reduce(pool.aditional_info.effects, pool, fn effect, pool ->
-          apply_effect_mechanic(pool, effect, game_state)
-        end)
-
-      put_in(game_state, [:pools, pool.id], pool)
-    end)
-  end
-
-  def apply_effect_mechanic(player, effect, game_state) do
-    now = System.monotonic_time(:millisecond)
-
-    Enum.reduce(effect.effect_mechanics, player, fn {mechanic_name, mechanic_params} = mechanic, player ->
-      should_re_apply? =
-        is_nil(Map.get(mechanic_params, :last_application_time)) or
-          (not is_nil(Map.get(mechanic_params, :effect_delay_ms)) and
-             now - Map.get(mechanic_params, :last_application_time) >= mechanic_params.effect_delay_ms)
-
-      if should_re_apply? do
-        do_effect_mechanics(game_state, player, effect, mechanic)
-        |> Effect.put_in_effect(effect, [:effect_mechanics, mechanic_name, :last_application_time], now)
-      else
-        player
-      end
-    end)
-  end
-
-  defp do_effect_mechanics(game_state, player, effect, {:pull, pull_params}) do
-    case Map.get(game_state.pools, effect.owner_id) do
-      nil ->
-        player
-
-      %{position: pool_position} when pool_position == player.position ->
-        player
-
-      pool ->
-        if player.aditional_info.damage_immunity do
-          player
-        else
-          direction = Physics.get_direction_from_positions(player.position, pool.position)
-
-          pull_foce =
-            pull_params.force + pull_params.force * pool.aditional_info.stat_multiplier
-
-          Physics.move_entity_to_direction(player, direction, pull_foce)
-          |> Map.put(:aditional_info, player.aditional_info)
-          |> Map.put(:collides_with, player.collides_with)
-        end
-    end
-  end
-
-  defp do_effect_mechanics(game_state, player, effect, {:damage, damage_params}) do
-    # TODO not all effects may come from pools entities, maybe we should update this when we implement other skills that
-    # applies this effect
-    Map.get(game_state.pools, effect.owner_id)
-    |> case do
-      nil ->
-        player
-
-      pool ->
-        pool_owner = Map.get(game_state.players, pool.aditional_info.owner_id)
-        real_damage = Player.calculate_real_damage(pool_owner, damage_params.damage)
-
-        send(self(), {:damage_done, pool_owner.id, real_damage})
-
-        player = Player.take_damage(player, real_damage)
-
-        unless Player.alive?(player) do
-          send(self(), {:to_killfeed, pool_owner.id, player.id})
-        end
-
-        player
-    end
-  end
-
-  defp do_effect_mechanics(game_state, player, _effect, {:buff_pool, buff_attributes}) do
-    Map.get(game_state.pools, player.id)
-    |> case do
-      nil ->
-        player
-
-      pool ->
-        update_in(pool, [:aditional_info, :stat_multiplier], fn
-          current_multiplier when current_multiplier > 0 ->
-            current_multiplier + current_multiplier * buff_attributes.stat_multiplier
-
-          _current_multiplier ->
-            buff_attributes.stat_multiplier
-        end)
-        |> update_in([:aditional_info, :duration_ms], fn current_duration ->
-          current_duration + buff_attributes.additive_duration_add_ms
-        end)
-    end
-  end
-
-  ## Sink for mechanics that don't do anything
-  defp do_effect_mechanics(_game_state, player, _effect, _mechanic) do
-    player
-  end
-
   defp calculate_angle_directions(amount, angle_between, base_direction) do
     base_direction_normalized = Utils.normalize(base_direction)
     middle = if rem(amount, 2) == 1, do: [base_direction_normalized], else: []
